Cleaned up instruments server and client api #85
