+++ cpp/include/krypto/instruments/exchanges/coinbase.h
@@ -8,9 +8,12 @@
 #include <nlohmann/json.hpp>
 #include <krypto/utils/types.h>
 #include <krypto/exchanges/coinbase/api.h>
-#include <krypto/instruments/exchanges/parser.h>
 
 namespace krypto::instruments::exchanges {
+    std::vector<krypto::utils::Instrument> parse(
+            nlohmann::json insts_json,
+            const std::unordered_map<std::string, krypto::serialization::Currency> &currency_reference);
+
     class CoinbaseInstruments : public ExchangeInstruments<CoinbaseInstruments> {
     private:
         krypto::exchanges::coinbase::PublicApi api_;
@@ -21,18 +24,4 @@ namespace krypto::instruments::exchanges {
 
         static const krypto::serialization::Exchange key = krypto::serialization::Exchange::Exchange_COINBASE;
     };
-
-    CoinbaseInstruments::CoinbaseInstruments(const Config &config, const std::string &environment) :
-            CoinbaseInstruments::ExchangeInstruments(),
-            api_{config, environment} {}
-
-    std::vector<krypto::utils::Instrument> CoinbaseInstruments::read_instruments() {
-        auto insts = api_.get_products();
-        if (!insts.has_value()) {
-            return {};
-        }
-
-        return krypto::instruments::exchanges::parse(insts.value(), CoinbaseInstruments::currency_reference);
-    }
-
 }
