+++ cpp/src/apps/instruments/server_main.cpp
@@ -26,7 +26,7 @@ int main(int argc, char **argv) {
 
     zmq::context_t context{1};
 
-    krypto::instruments::Server<true> server{context, config};
+    krypto::instruments::Server server{context, config};
     auto done = std::async(std::launch::async, [&server]() { server.start(); });
 
     shutdown_handler = [&server](int signal) {
